refactor pages files
